Rename EMR status checker file
